+++ app/src/main/java/com/ar/arstoken/data/db/CustomerDueRow.kt
@@ -2,5 +2,5 @@ package com.ar.arstoken.data.db
 
 data class CustomerDueRow(
     val customerId: Int,
-    val due: Int
+    val due: Double
 )
