+++ app/src/main/java/com/ar/arstoken/model/CartItem.kt
@@ -2,5 +2,5 @@ package com.ar.arstoken.model
 
 data class CartItem(
     val item: Item,
-    val qty: Int
+    val qty: Double
 )
