+++ app/src/main/java/com/ar/arstoken/model/Customer.kt
@@ -5,5 +5,5 @@ data class Customer(
     val cloudId: String,
     val name: String,
     val phone: String,
-    val creditBalance: Int = 0
+    val creditBalance: Double = 0.0
 )
